Add simple tests
Add test for input value conversion in both directions and test some
field validations

The schema.json is an GraphQL introspection result from the current test
project.

address.json a "DomainTypeInput" value, the address domainql type of that
example.

address-values.json is the same in it's internal formik form. All fields
are strings except for boolean which are boolean