Add withForm() HOC
Add a new HOC as an easy way to wrap a form component in a domainql form
iwith formConfig as prop etc.

```js
// in ForForm.js

class FooForm extends React.Component
{
    ...
}

export default withForm(
    FooForm,
    {
        type: "FooInput"
    }
);

```

The HOC takes two arguments, first the Component to be wrapped, and a
props object with props for the form component.

In FooForm, you can define the contents of the form, have formConfig
available as prop, and on the invoking side you can just invoke it
with

```js
<FooForm
    key={ fooDetail.id }
    value={ fooDetail }
    onSubmit={ this.onSubmit }
/>

```

by declaring just the starting value as `value` and the submit handler
as `onSubmit` prop.