Conversion: Handle undefined roots/formValues
If given a falsy root object, the conversion value would just throw with
some cryptic error in convertInput() which is not protected against
null/undefined values.

At first look, we could just start with convertValue() instead, which is
protected, but it just doesn't make sense to convert and edit totally
undefined values.

So we throw an exception in both of those cases and write a test
for those messages.