GlobalErrors: Fix path, render path as data-path
First of all, there were some errors in path creation and form passing.

Then we have to deal with the situation of a complex object not actually
fulfilling the restrictions its type puts on it.

There errors can occur even though we don't edit the fields and can be
quite puzzling.

To help debugging, we render a data-path attribute for fields that have
no htmlFor attribute, so you can see the target in the DOM inspector
in any case.