Conversion: don't provide empty containers.
So far, we have converted falsy non-scalar containers like array or
objects as [] and {} which just seems wrong in the end.

Yes, it is inconvenient to provide empty arrays, but it is better than
to brush over that and interpret an array from nothing.

Even moreso for objects because of missing properties.

We change that behavior and correct the tests as far as they're
concerned.

If you say for form works with "InputType", you need to provide valid
"InputType" objects.