Add buttonStatus option
Add an option "buttonStatus" (enabled by default) which makes dealing
with multiple submit buttons in a form easier.

It registers a click handler that watches for click on submit buttons
with name attribute. If the name attribute is set, the name is
set in the formik status object as "button" attribute.

We also add an optional onClick handler to the form.