Implement <AutoSubmit/>
Implements a signal component <AutoSubmit/> that can be dropped into a
domainql-form.

It renders nothing but listens for form updates. If the form has no
errors and the last form content is not deep equal to the next form
content, the component triggers a debounced submit for the form.

The debounce timeout can be set by calling
     <AutoSubmit timeout={ 200}/>

The default is 300ms.