Updgrade test to async
In the previous version of formik, our synch tests worked fine, now they
don't.

We opt for a strategy of using setImmediate to wait for the update to
finish on the theory that our setImmediate will be executed after the
ones in the tested code.

This might sound not too stable, but it works for now and seems
preferable to using setTimeout and run into timing problems.

This should be more deterministic no matter how fast the current machine
or how big the background load.