Add missing mode support to Select and TextArea
The <Select/> and <TextArea/> components should react to the surrounding
field context just like the default field does.

If the their own mode or a inherited parent mode is FieldMode.DISABLED
they need to render the input element as disabled.

FieldMode.READ_ONLY correspondingly causes a text-only rendering, in
this case we use the preexisting static rendering infrastructure.

Adds some tests for the new behavior.