Rewrite Tests for 2.0
The tests once worked for the formik version but were totally broken
since then.

This repairs all tests to work with mobx observables and fixes a lot of
issues that fell under the table.

This changes the Testing API from "enzyme" to "react-testing-library"
which provides everything we really need here and also seems to produce
more clean tests in the end.

We also use "mocha-param" and "cartesian" to run parametrized tests
testing all permutations of field modes and field mode specification
location (directly on the field vs inherited).

FormList and FormSelector do not work at this point but we retire them
for now.

For the mobx based forms with their default auto-commit on submit it
seems to make more sense to use multiple forms for individual
object-leafs of a larger data-graph than to use these fixed list and
selector contructs.

Maybe we'll readd something similar later.

High-level validation is still lacking here since we need to decide on a
general extension interface that can work in general and also works in
providing the complicated stuff we need for LIS-A.

Minimized Schema
----------------

We centralize schema access to implement a class "TrackingSchema" that
extends InputSchema and tracks the types requested from it in a set.

We add a little util function test/util/dumpUsage.js that we call with
after() whenever we use the centralized schema.

The function which is commented out by default updates a file
"types-used.json" that contains all the object types actually accessed
during the tests.

We use this knowledge to minimize the humonguous test schema to just
what we use for the tests.