Implement high-level validation
For LIS-A we need very specialized validation with user-admin editable
validation rules etc pp, but it seems a very bad idea to push these very
specific functionality into domainql-form. We might never used it again
or not like this and there are many many other ways to implement
high-level validation.

Validation Interface
--------------------

This implements a simple hook-system with which external validators can
plug themselves into domainql-form. The interface is very simple and
based on the internal structure of domainql with low overhead.

It adds a validation prop to <Form/> or any other of the form
configuration blocks. This validation prop receives an object with named
hooks.

```js

    <Form
        type="MyDomainType"
        validation={
            // Validation Object
            {
                fieldContext: ctx => { ... }
                validateField: (ctx, value) => { ... }
            }
        }
    />

```

`fieldContext(ctx)` allows the validator to modify the field contexts of fields
to either change internally used values or to add data to the field
context to cache field resolution results.

`validateField(ctx,value)` is doing the actual high-level field
validation. `ctx` is the field context for the field, `value` is
the current field value. The high-level validation will only be called if the
value matches the type validation criteria of the field but is still the
original value of the field.

Using named hooks we can add other validation hooks later.