Internal API refactoring
Working with the high-level validation it became clear that we were
recreating the fieldContext far too often. It should only change for
very basic changes on the field. The field being disabled, classes
changing on the field and not on every key stroke as it now was.

Fixing that however revealed another decision to be very stupid which is
moving the formConfig into the field context which necessitates changing
the field context every time to reflect changes on fields that trigger
errors. Since we don't want to write back these erroneous values into
the mobx view model, we keep the current field value as first error
value, which we keep *in the form config*

This commit changes the internal Field renderers but also the "field
with render function child" use-case to not only receive two arguments
(formConfig, fieldContext). First the current form config and then the
field context that now doesn't contain the form config.

This commit also changes the Field state handling to minimize the the
field context changes to a sensible set of field attribute changes.

Some changes just should never happen. Fuck off with fields dynamically
changing their data-graph positions and types.

Changing the label seems very unlikely now, too, but we might reactivate
that later.