Refactor domainql to use hooks
Since we're still so early in our development cycle, we might just as
well do it right.

We move the complete form element implementation over to a hook based
implementation which in many places seems nicer than before and also
allows us more fine-granular updates on form configuration and errors.

All the ports are pretty straight-forward, in some cases we can
transfer getDerivedStateFromProps to a useMemo() in some cases with and
added useCallBack(getDerivedStateFromProps to a useMemo() in some cases
with and added useCallBack(). GlobalErrors uses a useEffect() to update
the cross-linking between global errors and form fields on first render
and later.

We retire our withFormConfig HOC. Class-based React components can still
use the FormConfigContext to receive the context, we will assume
functional components with hooks from now on.

We add a maybe silly useFormConfig() hook to our library which accesses
the context for the current form config and disuades the users from
caring about aforementioned FormConfigContext for the recommened usage.