Fix non-null check
During the porting from formik to the current system the non-null checks
for fields got lost along the way and there never was an actual test.

Adds a test case for a required field.

Refactoring
-----------

We make the local handling of the errors plus value less awkward by
defining an existing list with the value and then just pushing/concating
into that list.

This requires some changes to later condition checks