Introduce internal context for Form
With the latest changes the signature of setFormContext has been
expanded and it turns out it's still not enough because we're missing a
way to programmatically trigger a form submit from a formConfig now.

This introduces a simple InternalContext object that contains the
secondary, mostly not for normal use parts of our formConfig.

We move functors to internal form concerns into it and leave the rest
of the parameters as is.

As of now these are:

 * setRoot - to set a new root object (not used in the code so far)
 * setErrors - to set a new errors state
 * handleSubmit - the original Form submit handler to be triggered
                  programmatically

The new internal context allows us to evolve these rarely used internal
helpers in a central place.

We store the internal context as formConfig.ctx.