Add tests for WireFormat
We were lacking tests for our WireFormat class that converts back and
forth between GraphQL data as it goes "over the wire" as JSON and the
same data as live JavaScript objects.

For example a `Timestamp` value is a ISO-8601 string while as live
javascript object it is a Date.

Over time we will add more and more complex conversions.