Implement conversion of aliased data
GraphQL queries can rename fields on the fly which on one hand of course
breaks the underlying type as the rules no longer apply, but at the same
time one could and I think should expect the scalar conversions to be
run nevertheless.

This commit adds two optional parameters to WireFormat.convert().

The first is a map of aliases

    {
        "foos.timestamp" : "f2"
    }

The unaliased qualfied name "foos.timestamp" within the object to be
converted is called "f2" in the data to be converted.

The conversion will read the value of the field `f2` and convert it as
if it was the original `timestamp` as specified in the type.

If a class implementation is given for the type, it will *not* be used
as the renamed types break the expected object shape (would e.g. add
unobserved fields to mobx objects).