Add support for dynamic forms
So far, we've only been working based on fixed types described in the
GraphQL schema. It can however be really useful to be able to create
ad-hoc forms based on dynamic data for example.

This commits refactors the internal type handling for fields so that it
uses a new method `formConfig.resolveType(name)` that by default does
the normal type resolution via InputSchema.

This commits then allows defining a GraphQL schema type structure and
specifying that via `<Form type={ typeStruct } />` as alternative of
giving a type reference by name.

This commits also adds number of simple factory methods that produce the
parts of GraphQL type definitions and which can be used to create the
necessary type structures.

Example from the test:

    const adHocType = inputObjectType(
        "AdHocType",
        [
            inputField(
                "field1", nonNull( scalar("String"))
            ),
            inputField(
                "field2", scalar("Int")
            )
        ]
    );

The fields of the new type must must be of a scalar or object
type that already exists within the schema.