Support inline forms
This commit adds support for inline forms and refactors the way form
layout is configured.

So far we only had a `horizontal` flag option, which now changes to a
String based enum FormLayout

    const FormLayout = {
        DEFAULT : "DEFAULT",
        HORIZONTAL : "HORIZONTAL",
        INLINE : "INLINE",
    };

The inline mode sets "form-inline" on the form element, suppresses the
rendering of form-group elements and adds a form-check-inline to
checkboxes.