<AutoSubmit/> is dead, long live autoSubmit
The very hacky nature of our <AutoSubmit/> led to it just not working
anymore, no matter how renovated.

AutoSubmit is not rerendered, we can't observe() on our view model
either.

The easiest way to implement seems to draw the feature into Form and
make it happen with an additional debounced submit callback.

There are two submit callbacks now, which is kind of unfortunate, but on
the other hand, we don't want to introduce a mandatory form submission
delay while the one submit is waiting for whether there is another one.
Using another implementation with leading edge debouncing doesn't seem
good either, since we mostly do want late submission when we do lazy
auto submit.

So we add another debounced submit handler to Form which gets triggered
every time the form changes without error.

We control the behavior with the two new Form options. Setting
"autoSubmit" to true enables the new behavior which is disabled by
default.

The timeout for the debounced submit handler is controlled by the
"submitTimeout" option, default 500.

getOption(name,options,parentOptions)
=====================================

For the submitTimeout, we need an somewhat akward way to get
the real merged option before we have a formConfig and without
introducing a cyclic input dependency on it.

So we add a new getOption helper method.