Redo dynamic forms
The way to handle dynamic forms we introduced with 671ca49 is overly
cumbersome in requiring to set up a formal model for the ad-hoc type.

Instead we just allow leaving out the type prop on <Form/> and define
the scalar type on the <Field/>s instead.

We support the Schema syntax for non-null fields (e.g. "String!") and
generate a "NONE.field:Field is required" error if is null.

Adds a test for dynamic forms.

This reverts commit 671ca49a2de4b4d111453f4332186b8eb28b6bc9.