Select: Add support for boolean and number values
Enable the name/value values prop to contain boolean and numeric values
and have those be written through to both the view model as well as
submit them to the observable.

Bugfix in InputSchema
=====================

There was a left-over clause that would replace all falsy results of
InputSchema.scalarToValue() with "".

But we want true booleans for booleans, so we scrap it.