Add GlobalConfig.valueOrNone(value)
Since we're already storing what to write for "none", let's also add a
standard method of substituting those none values.

"none" should be substituted for every empty value ( null or "") but we
also want to allow undefined as this can be useful sometimes.

So all these values are displayed as none() value and all others are
returned as strings as-is, including falsy values like 0 or false.