Implement Field addons for <Field/> and <TextArea/>
This commits adds support for Bootstrap 4 addons for domainql-form. The
addons are placed inside the <Field/> or <TextArea/> elements to render
that field with an addon.

    <Field name="price">
        <Addon placement={ Addon.RIGHT }>
            EUR
        </Addon>
    </Field>

There's also the possibility to give <Field/> and addons array property.
This is mostly used internally for fields like <TextArea/> that use a
render function form of <Field/> internally.

Note that the support of Field for addons of course only lies within the
default-renderer. If you change that default-renderer, you will need to
do it yourself with the help of two static helper methods.

Addon.filterAddons(children) filters a react children object into an
array that contains only Addon elements.

Addon.renderWithAddons(fieldElement, addons) renders a fieldElement with
the given addons or just a field element if there are no addons.

Adds two tests.