FormConfig: Add programmatic error handling
So far all our validation errors where the result of the automatic
scalar based type conversion or the additional high-level validation.

But there are still cases where we'd like to be able to add errors to
form elements without having such a validation.

This adds two methods `formConfig.addError(path,msg,value)` and
`formConfig.removeErrors(path)` to allow programmatic control of errors
for any participating component of a Form.

The validation code already does per-path-prop updates of the error list
so it already works with this changes.

Note that the "path" object of an error entry does not actually need to
coincide with an actual path within `formConfig.root` if you don't use
the `formConfig.getValue(path, errorMessages)` method which would try to
fetch a value from the path if there are no errors.

You can use any arbitrary unique path prop for your error entries if you
make sure that your input elements render the same value as `name` prop.
This is necessary for <GlobalErrors/> to directly link to your component
in the error labels it generates.