Addon: Tweak observable
We keep having persistent problems with the observable content given in
an <Addon> does not update as it should.

The children object given into Addon should have been produced in a
proper observer context inside the Form implementation, but it just
doesn't update.

I tried tweaking the observer-setup of multiple components but could not
get to a satisfactory result.

What is left is evading the issue by allowing a <Addon/> to receive a
single function child that gets evaluated within the observer context
of the Addon component itself.

This wraps <Addon/> in its own observer and handles the rendering of
function children.