Refactor Library (breaking change)
The current state of the 2.x versions of domainql-form still had some
issues in terms of form isolation. The mobx view model would provide
some form input isolation, but only on the root level. Furthermore the
more I thought about it the clearer it seemed that even the existing
isolation feature was a very awkward fit into the way we do things in
automaton.

Sure, for a simple flat object the <Form/> component would isolate the
user-input during form input, but then that isolation would end in a
very awkward way by first writing back / committing the current values
to the original observables to then invoke a transition function that
receives the same object as context and does what we want to do with the
form object after editing.

I think what we really want is a totally isolated, observable containing
the current form input which we then can process without interferring
with the original observable or replace the original observable with the
new content or something totally different.

To handle such cases, I had already written a clone function for domain
objects that can create deep clones of domain object graphs.

This commit now changes the general form behavior in such a way that it
doesn't use view-models anymore, but which clones the original
observable as internal root object. Optionally this cloning can be
skipped, so that the form works directly on the original observable,
which can be very useful for search forms etc.

This commit integrates the clone module from automaton into InputSchema,
creating two methods InputSchema.clone() and InputSchema.cloneList().

"isolation" option
------------------

A new option "isolation" controls the cloning of the root object. It is
set to `true` by default, isolating the object changed by the form by
cloning it before editing.

Setting isolation to `false` will skip the cloning and the Form will
work on the original observable. All changes will be applied immediately
without debouncing.

The original autoSubmit feature is still in place, so you can also keep
isolation enabled and just setup the debounced auto-commit as before.

Fallback clone
--------------

For forms without formal GraphQL-type, we need a third clone function,
that can clone normal JavaScript objects. The default implemenetation
works for now, but is not optimal in that it might still confuse scalar
values embedded within an un-typed object with an an object to be cloned
as object. We already special-case `Date` to prevent that but there
might be more cases, especially considering the extensible scalar value
types.

Domain Object Factory / Fall-clone customization
------------------------------------------------

The library also exports two customization functions

    registerDomainObjectFactory()

    Registers an alternate function to create new objects while cloning
    domain object graphs.

    registerFallbackCloneFunction()

    Registers an alternate JS object clone function for non-type form
    object root objects

Breaking Changes
----------------

The actual changes required where rather small, but as they are causing
fundamentally different behavior, this is a breaking change, even though
it might not actually change observable behavior in applications.

This commit changes a lot of asserts in the tests asserting the original
loss of isolation and which are now asserting the perpetual isolation by
default.