Refactor Icon
The <Icon/> component was both too limited and too much geared to our
own current usage for LIS-A.

This commit gets rid of the `brand` prop and replaces it with a general
`family` prop that takes a symbolic name and converts it to the
corresponding fontawesome class.

We export the valid values as static members of Icon

 * Icon.SOLID
 * Icon.REGULAR
 * Icon.LIGHT
 * Icon.DUOTONE
 * Icon.BRAND

These are all the families fontawesome currently has. Note that you need
to provide the corresponding stylesheets in your base template / CSS.

So a brand icon is now defined as

    <Icon
        className="quinscape"
        family={ Icon.BRAND }
    />