Clone: Ensure undefined id remains undefined
Makes sure that if an object that is cloned and has a an undefined id,
the clone has an undefined id, too.

This was pretty much already given for every fields but id, but due to
our domain object creation process, the id would end up as null.

This causes subsequent conversion operations of that partial object to
error out since it will ignore an undefined id but will error on a
defined id that is null.

Adds a test.

There is implicit clone testing within the form tests, so this
adds just a test case for the special error case we fix here.