Breaking: Implement ErrorStorage
So far the <Form/> component has been keeping the errors as local state
and passed control of them over to FormConfig.

I initially envisioned there only being one form for an object, but for
lots of layout reasons it often would be preferable to split that into
multiple forms.

With turned-off isolation this works out fine, the only thing missing is
error handling.

This commit changes the way that the errors are stored.

It introduces a new class ErrorStorage that uses a WeakMap to store the
errors for each observable.

This means that forms sharing the same form root object are now also
sharing the same error context, that is their buttons react the same etc
pp.

We allow passing in the errorStorage to use as third constructor
parameter of FormConfig and default to a static instance if none is
given.

An errorStorage prop can be passed to both <Form/> and <FormConfig/>
with the latter being the recommended version.

Adds a test

Since this commit changes the behavior of the library it is considered a
breaking change, however it does not require you to change anything if
you do not, for some reason, rely on separate errors for your forms.

Since this commit changes the behavior of the library it is considered a
breaking change, however it does not require you to change anything if
you do not, for some reason, rely on separate errors for your forms.