<Field/>: Offer onChange / validate
Enable local field onChange handler and local validation functions.

    <Field onChange={ (fieldContext, value) => ... } />

The onChange handler is called whenever the input field changes.

    <Field validate={ (fieldContext, value) => ... } />

The validate handler is called to check the validity of every new value
within a <Field/>. The handler can return an error message string or
`null` to signal success.

First the type validation happens. If that succeeds, then the local
validation happens. If that succeeds, the high-level validation is
called.

A failure of the local validation will prevent the corresponding value
to be written back into the observable.