Add Addon.BEFORE and Addon.AFTER
Addons are our primary mechanism to customize the rendering of <Field/>
components and components using <Field/>.

They are born out of the addon concept in bootstrap for which it only
means that the addon is visually presented as part of the visual input
field.

Our <Field/> abstraction however encapsulates the whole form-group in
the end and doesn't provide for means to add stuff to that form group.

This commit extends our Addon abstraction with two new placements:
BEFORE and AFTER. They are not addons in the boostrap sense but are just
wrapper for elements that are rendered right before or right after the
field element or input group containing the field element is rendered.

The changes are necessary only within the Addon module.

Refactors the way we render all addons. Makes sure to not render
superfluous input groups for fields that have only BEFORE and/or AFTER
addons.

Adds two test cases for BEFORE and AFTER addons.