Addon: Add moveIfPlainText prop
Adds a new prop to the Addon component called moveIfPlainText. Setting
this prop to true allows to place an addon LEFT or RIGHT but then have
it moved to BEFORE/AFTER if the corresponding field is set to plain
text.

Changes the default renderers to use Addon.renderWithAddons() in any
case. Custom field implementations are advised to do the same to enable
rendering of BEFORE/AFTER addons in any case, moved or not.

Adds a test case.