Fix GlobalErrors:
Global errors component wasn't updating right anymore because while it
was retrieving the global errors it wasn't getting the updates it needed
to actually work.

We need to make sure to read the length of the observable errors array
to make sure we're notified of changes. We use the received value as
data attribute to make sure this isn't optimized away.