Implement revalidation
So far we've been relying on change events to make sure our data is
valid, assuming that all data is initially in a valid state.

However when creating new objects it can be useful to be able to create
an object with empty fields even though those fields are finally NOT
NULL.

So when someone doesn't fill in all values, we actually never get a
change event for those fields and so far did not recognize the error.

This would make the erroneous data run into the next GraphQL endpoint
which is of course far from optimal.

This commit introduces a revalidation function which is invoked when
submitting by default (set the new revalidate option to false to
disable this).

It recurses over all GraphQL fields within the root object and validates
all values that have corresponding fields in the current mask / error
storage context.

It then performs NON_NULL and high-level validation for each of those
fields and adds an error if the field is invalid.

The form does not submit when there were errors.

FieldContext registration
=========================

For this functionality we need to be able to access a field context from
the outside. So we start to register every field context for the root
object we're creating it for.

Field.registerContext(root,fieldContext) registers the field context.
This is normally an internal function executed by the field context
creating within <Field/>

Field.lookupContexts(root) returns a Map mapping the qualified name of a
field context (in non-GraphQL conventions (e.g. "rows.0.name")) to the
fieldContext for that qualified name.