<Field/> / <TextArea/>: Add maxLength prop
Adds a maxLength prop to <Field/> and <TextArea/> that validates the
maximum length of the value for "String" typed fields.

We always use a soft limit that produces a "Value Too Long" error when
the value is too long. We do not limit the actual value entered by
the user.

Adds two tests.