Improve static booleans a11y
The static rendering of booleans is just an icon without text, so we add
text "true"/"false" both as tooltip as well as aria-label

Keep the green checkmark but don't rely on red/green distinction and
also avoid representing "false" as danger.

Instead we just use text-muted to get a dark grey color for our
checkmark less box.