Export resolveStaticRenderer
Export the method to resolve a static renderer from the library and give
it a option object with one "default" option so far that controls
whether the method returns the default renderer if no specific renderer
is registered.

This allows checking from the outside if a certain renderer is already
registered.

e.g. we register a static renderer for Timestamp/Date in automaton
but only if the user hasn't registered an alternative in the
app-startup.js init function.