Tolerate null form values
Tolerate passing in a null value for a form object and disable all form
elements within that form.

Make sure that we remain vigilant towards wrong field names that are not
according to GraphQL spec, but on the other hand, we're tolerant on
reading non-existing values for fields.

We already kind of tolerated that because the underlying lodash function
is very tolerant and accepts both get(null, "bla") as well as
get(obj, "nonExistingProp"). But it returns "undefined" which the inputs
don't like as value.

So we make sure to substitute that undefined with null.

Adds tests.