Refactor FormContext
Recently some assumptions I had about the best way to do forms with
domainql-form turned out to be not what we actually want to do and what
I think is the better way to handle it.

When I intially started I imagined that there would be a single
potentially big form corresponding to a potentially big object graph
where we address all fields relative to the one root object of that
hierarchy.

But first of all it turned out to be easier to have many different forms
within a view for layout purposes.

The FormBlock I initially envisioned to fullfil that role doesn't always work
because other elements contain forms, too, so we can't put them inside a
<Form/> since we can't nest forms in HTML.

So now we have one <Form/> followed by e.g. a datagrid with filter form,
followed by another <Form/>.

Also it is often cumbersome to handle subobject forms and create the
correct qualified paths etc.

It is much easier to have e.g. a form for the root object and then
maybe another for the nTh sub-object of within a list field.

And to make things more complicated, it should be possible that multiple
forms and forms for sub-objects or even completey different objects are
logically within one logical form context and present themselves as
such.

Any existing error within that logical context should disable any
non-discarding function within the same context. (We can do a "Cancel"
from an erroneous state, but not a "Save", duh)

The existing ErrorStorage kinda went in the right direction but in the
end did not fullfil all requirements.

This commit refactors the former ErrorStorage into a new class called
FormContext.

The FormContext stores all errors within that context and can provide
all errors or errors for a specific root object.

It contains the formerly static field registration which allows for easy
iteration over all existing form fields within that FormContext for
revalidation.

The revalidation gets refactored to actually validate all form fields
within the context if any of its forms get submitted.

We delegate some of the field change handling from FormConfig to
FormContext so that only FormContext needs to know about its internal
storage format.

Internally, we use a WeakMap to provide and remember unique numeric ids
for observable objects without preventing the garbage collection.

This cuts down our two uses of WeakMap to one, eliminating a
questionable use where we had a WeakMap for internal objects. Now the
fieldContexts are just stored in the FormContext.