Fix form context edge cases
The initial refactoring failed to account for the fact that we have
situations where we access a formConfig and it's there because the
<FormConfigProvider/> provides it, but it is outside of a form so it has
no form context.

We already had that situation before and kind of handled it, but now we
can do so more cleanly because we can disambiguate between a missing
formContext meaning we're outside of a form and never have errors and a
present formContext without form Object.

This adds tests to that effect.

Secures the FormConfig interface for missing formContexts. FormConfig
reports there being no errors always, methods to add errors throw on
missing form context.