InputSchema: Fix resolve bug
There were a lot of implicit tests for the resolveType method of input
schema, but no explicit ones. Turns out the implicit ones still had a
broken edge case when the paths get longer.

This adds an explicit test to test several resolveType variants.

Note: I stopped using the "getSchema()" mechanism still used in many
tests here.

It was a work-around the fact that I used very large schemas imported
from automaton-test and wanted to reduce them for the tests, so the
solution was to track what the test accesses and then to reduce the
schema to that.

It kinda worked, but there's a much better approach with the
pickSchemaTypes helper which allows us to generate a minimized schema
from a larger schema. We take a huge schema and specify an array of
types we're interested in and pickSchemaTypes() will follow the type
references of those types and create a minimal schema with referential
integrity.