Field: Allow for tooltip changes
I initially thought of the tooltip as unchanging, but now we have cases
where the client has catalog values that are too long the editing UI so
they want the current value as tooltip.

This enables that by adding `tooltip` to the field context memo inputs
where it was missing.