Refactor high-level validation and schema handling
This commit concludes the logic of the former form context relating
commits.

If we manage many forms it makes no sense for these forms to have
different validation rules, *unless they have*, but then it is clearly
better to isolate them in different form contexts.

Schema has always been the odd one out among the context propagated
values and this commit draws it all out into form context.

The constructor of FormContext now requires the schema instance and
receives an additional options object with the validation moved over
from <Form/> being the only option for now.

    new FormContext(inputSchema, { validation: ... })

We provide a better API for setting and getting the default form
context.

    FormContext.getDefault()

returns the globally registered default form context and calling

    formContext.useAsDefault()

on a form context turns that form context into the default form context.

The default form context is the one that is used by all formConfigs and
all <Form/>s unless another one is specified.

This adjusts the test cases.

"Outside"
=========

This removes the concept of "outside" the form. The formConfig outside a
form is connected to the same form context as the one inside the form.