Field: Reserve section in context
We initialize the a "section" property to null in our shared field
context, to allow automaton or other users to attach a section
identifier to the field from outside while making sure future code will
know that there's a section prop already.

Not optimal but the best we can do in this situation.