Upgrade dependencies
We had accumulated quite some upgrade debt, so the upgrade was a bit
rough.

Adds a lot of changes to tests to account for changes in behavior in the
upgraded libs.

We create <span> elements pointed to by labels at some points, the old
version of the lib used to tolerate that.

Adds findSpanByLabelText(container,text) to account for those cases.

The new userEvent.type() just types the argument after the contents and
requires a special selection statement to preselect and thus delete the
former content

Adds clearAndType(container,text) to clear an input and type the given
text.

In some cases we need to call makeObservable() on some of our classes
because, while they are not observables, they contain @action
decorators we need to activate that way. Since we have no @observable
decorators we have no observable properties.