Change empty value validation
By convention, we do not want to distinguish between an empty string and
a null value. The user entering an empty string should always be
interpreted as a null value.

Going by that, we had a very early shortcut in the validation that
return "no error" early, which prevented the empty string values from
showing up in the type based validation and later high level validation.

For the type based validation, that's ok. If the value is null, we don't
need a type-based null-value validation, that concern should be handled
by the required/nonnull handling.

For the field-level and high-level validation however, we might want to
declare optional required fields that hinge on another field having a
certain value. If you define this, you also need to define that. Things
like that.

This changes validation so that an empty string only skips type-based
validation but then gets passed on to local or high-level validation.

Adds a test, extends a test.